seismology/web/requirements.txt
flask
Flask-Breadcrumbs
python-dotenv
requests
